test2

+++ .github/release.sh
@@ -1,8 +1,16 @@
 #!/usr/bin/env sh
 set -o errexit
 
-echo "//npm.pkg.github.com/:_authToken=$1" > $HOME/.npmrc
+if [ "$GITHUB_ACTIONS" = "true" ]; then
+  echo "//npm.pkg.github.com/:_authToken=$1" > $HOME/.npmrc
+else
+  echo "//npm.pkg.github.com/:_authToken=$1" >> .npmrc
+fi
+
 yarn workspace @radist2s/app run build
 yarn workspace @radist2s/app run build:paths-dependencies
 yarn changeset publish
-(git checkout HEAD -- .npmrc 2> /dev/null) || (rm .npmrc 2> /dev/null) || true
+
+if [ "$GITHUB_ACTIONS" != "true" ]; then
+  (git checkout HEAD -- .npmrc 2> /dev/null) || (rm .npmrc 2> /dev/null) || true
+fi
